fix playground preview css
